Promiseでラップした関数を別ファイルに切り出し
